+++ lib/regexp_parser/expression/classes/posix_class.rb
@@ -1,9 +1,5 @@
 module Regexp::Expression
   class PosixClass < Regexp::Expression::Base
-    def negative?
-      type == :nonposixclass
-    end
-
     def name
       text[/\w+/]
     end