+++ lib/regexp_parser/expression/classes/unicode_property.rb
@@ -1,10 +1,6 @@
 module Regexp::Expression
   module UnicodeProperty
     class Base < Regexp::Expression::Base
-      def negative?
-        type == :nonproperty
-      end
-
       def name
         text[/\A\\[pP]\{([^}]+)\}\z/, 1]
       end